Code function for playing a single round
Just a very basic if/else if/else structure for deciding who wins during
each round. Both choices matching, i.e. a draw, was taken as the base
case, with further if/else statements to handle the other outcomes when
the user has picked rock, paper or scissors.

I'm not sure how useful the output is for coding the full-game function
however, so it probably needs some tweaking. Perhaps stripping the
scorring updates out of the playround function, and just let it alert
who won the round and return a single-word string of win, lose or draw
as appropriate.