Code function for a full game of 5 rounds
I opted for a simple while loop, checking against a variable that stores
the number of rounds played.

The play a round function was altered slightly since it is now
contained with the play game function. So it purely logs a message to
the console and returns a string of "win" "lose" or "draw".

Incrementing of the round number and scoring variables were all handled
within the while loop itself. And then some final if statements once the
while loop completes to check and report who actually won the game.